+++ packages/example-advanced/package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "accept-language-parser": "1.5.0",
-    "date-fns": "^2.16.1",
     "lodash": "^4.17.21",
     "next": "^13.0.0",
     "next-intl": "^2.10.4",
