+++ packages/use-intl/src/core/index.tsx
@@ -8,4 +8,7 @@ export {default as DateTimeFormatOptions} from './DateTimeFormatOptions';
 export {default as NumberFormatOptions} from './NumberFormatOptions';
 export {default as IntlError, IntlErrorCode} from './IntlError';
 export {default as createTranslator} from './createTranslator';
+export {default as createFormatter} from './createFormatter';
+
+// TODO: Remove in next major version
 export {default as createIntl} from './createIntl';
