+++ packages/use-intl/src/react/useIntl.tsx
@@ -2,8 +2,18 @@ import {useMemo} from 'react';
 import createIntl from '../core/createIntl';
 import useIntlContext from './useIntlContext';
 
+let hasWarned = false;
+
+/** @deprecated  Switch to `useFormatter` */
 export default function useIntl() {
   const {formats, locale, now: globalNow, onError, timeZone} = useIntlContext();
+
+  if (!hasWarned) {
+    hasWarned = true;
+    console.warn(
+      '`useIntl()` is deprecated and will be removed in the next major version. Please switch to `useFormatter()`.'
+    );
+  }
 
   return useMemo(
     () =>
